requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
black==24.4.2
bleach==6.1.0
boto3==1.34.111
botocore==1.34.111
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
cryptography==42.0.7
dask==2024.5.1
dask-expr==1.1.1
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.5.0
fuzzywuzzy==0.18.0
greenlet==3.0.3
huggingface-hub==0.23.2
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonformer==0.12.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.2.0
langchain-aws==0.1.5
langchain-community==0.2.0
langchain-core==0.2.1
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.0
langsmith==0.1.62
Levenshtein==0.25.1
locket==1.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pdfminer==20191125
pdfminer.six==20231228
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyenchant==3.2.2
pyflakes==3.2.0
Pygments==2.18.0
pypdf==4.2.0
pypdfium2==4.30.0
pytesseract==0.3.10
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.1
ratelimit==2.2.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-mock==1.12.1
rpds-py==0.19.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.4.4
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
stack-data==0.6.3
swifter==1.4.0
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
XlsxWriter==3.2.0
yarg==0.1.9
yarl==1.9.4
